Fix duplicate display tree issue
If the application hasn't received enough trees to fill its ring
buffer and the user pauses the view, then trees that don't get exist
will be paused to the pause buffer.  Later after the user selects play
to view the latest incoming tree set then there could be duplicate
trees (of the same tree) in the view.